dcgan-from-scratch/requirements.txt
pandas
lightning
torchmetrics
torchvision
tensorboard

black